refactor(spark): remove redundant SparkAdapter methods
Remove two SparkAdapter methods with no callers (createInterpretedPredicate,
sqlExecutionWithNewExecutionId) and de-duplicate methods whose bodies were
byte-identical across all per-version adapters:

- Trait defaults on SparkAdapter (identical across all six versions):
  isColumnarBatchRow, createCatalystMetadataForMetaField,
  extractDeleteCondition, convertStorageLevelToString.
- Pulled up to BaseSpark3Adapter / BaseSpark4Adapter (identical within a
  family): injectTableFunctions, injectScalarFunctions,
  injectPlannerStrategies, createLanceFileReader, and for 4.x also
  stopSparkContext, isLegacyBehaviorPolicy, isTimestampNTZType,
  getRebaseSpec, isHoodieTable.
- Spark3_3Adapter keeps its createLanceFileReader override returning None
  (no Lance support on Spark 3.3); all other removals are pure de-dup with
  no behavior change.
- injectTableFunctions becomes abstract on the trait: the no-op default was
  dead since both base adapters override it.
