+++ hudi-spark-datasource/hudi-spark4.2.x/src/main/scala/org/apache/spark/sql/adapter/Spark4_2Adapter.scala
@@ -26,7 +26,6 @@ import org.apache.hudi.common.table.cdc.HoodieCDCFileSplit
 import org.apache.hadoop.conf.Configuration
 import org.apache.parquet.schema.{GroupType, LogicalTypeAnnotation, Types}
 import org.apache.spark.SparkEnv
-import org.apache.spark.api.java.JavaSparkContext
 import org.apache.spark.sql._
 import org.apache.spark.sql.avro._
 import org.apache.spark.sql.catalyst.InternalRow
@@ -39,23 +38,16 @@ import org.apache.spark.sql.catalyst.planning.PhysicalOperation
 import org.apache.spark.sql.catalyst.plans.logical._
 import org.apache.spark.sql.catalyst.trees.Origin
 import org.apache.spark.sql.catalyst.types.DataTypeUtils
-import org.apache.spark.sql.catalyst.util.{METADATA_COL_ATTR_KEY, RebaseDateTime}
 import org.apache.spark.sql.connector.catalog.{V1Table, V2TableWithV1Fallback}
 import org.apache.spark.sql.execution.datasources._
-import org.apache.spark.sql.execution.datasources.lance.SparkLanceReaderBase
 import org.apache.spark.sql.execution.datasources.parquet.{ParquetFileFormat, Spark42LegacyHoodieParquetFileFormat, Spark42ParquetReader}
 import org.apache.spark.sql.execution.datasources.v2.DataSourceV2Relation
 import org.apache.spark.sql.execution.datasources.vortex.SparkVortexReaderBase
 import org.apache.spark.sql.execution.streaming.runtime.MemoryStream
 import org.apache.spark.sql.hudi.{HoodieMemoryStream, SparkAdapter}
-import org.apache.spark.sql.hudi.analysis.TableValuedFunctions
-import org.apache.spark.sql.hudi.blob.{BatchedBlobReaderStrategy, ScalarFunctions}
 import org.apache.spark.sql.internal.{LegacyBehaviorPolicy, SQLConf}
 import org.apache.spark.sql.parser.{HoodieExtendedParserInterface, HoodieSpark4_2ExtendedSqlParser}
-import org.apache.spark.sql.types.{DataType, DataTypes, Metadata, MetadataBuilder, StructField, StructType}
-import org.apache.spark.sql.vectorized.ColumnarBatchRow
-import org.apache.spark.storage.StorageLevel
-import org.apache.spark.storage.StorageLevel._
+import org.apache.spark.sql.types.{DataType, DataTypes, StructField, StructType}
 import org.apache.spark.unsafe.types.UTF8String
 
 import scala.jdk.CollectionConverters.MapHasAsScala
@@ -80,17 +72,6 @@ class Spark4_2Adapter extends BaseSpark4Adapter {
       }
     }
   }
-
-  def isHoodieTable(v2Table: V2TableWithV1Fallback): Boolean = {
-    v2Table.getClass.getName.contains("HoodieInternalV2Table")
-  }
-
-  override def isColumnarBatchRow(r: InternalRow): Boolean = r.isInstanceOf[ColumnarBatchRow]
-
-  def createCatalystMetadataForMetaField: Metadata =
-    new MetadataBuilder()
-      .putBoolean(METADATA_COL_ATTR_KEY, value = true)
-      .build()
 
   override def getCatalystExpressionUtils: HoodieCatalystExpressionUtils = HoodieSpark42CatalystExpressionUtils
 
@@ -136,44 +117,6 @@ class Spark4_2Adapter extends BaseSpark4Adapter {
     new Spark42HoodiePartitionFileSliceMapping(values, slices)
   }
 
-  override def extractDeleteCondition(deleteFromTable: Command): Expression = {
-    deleteFromTable.asInstanceOf[DeleteFromTable].condition
-  }
-
-  override def injectTableFunctions(extensions: SparkSessionExtensions): Unit = {
-    TableValuedFunctions.funcs.foreach(extensions.injectTableFunction)
-  }
-
-  override def injectScalarFunctions(extensions: SparkSessionExtensions): Unit = {
-    ScalarFunctions.funcs.foreach(extensions.injectFunction)
-  }
-
-  override def injectPlannerStrategies(extensions: SparkSessionExtensions): Unit = {
-    extensions.injectPlannerStrategy { session =>
-      BatchedBlobReaderStrategy(session)
-    }
-  }
-
-  /**
-   * Converts instance of [[StorageLevel]] to a corresponding string
-   */
-  override def convertStorageLevelToString(level: StorageLevel): String = level match {
-    case NONE => "NONE"
-    case DISK_ONLY => "DISK_ONLY"
-    case DISK_ONLY_2 => "DISK_ONLY_2"
-    case DISK_ONLY_3 => "DISK_ONLY_3"
-    case MEMORY_ONLY => "MEMORY_ONLY"
-    case MEMORY_ONLY_2 => "MEMORY_ONLY_2"
-    case MEMORY_ONLY_SER => "MEMORY_ONLY_SER"
-    case MEMORY_ONLY_SER_2 => "MEMORY_ONLY_SER_2"
-    case MEMORY_AND_DISK => "MEMORY_AND_DISK"
-    case MEMORY_AND_DISK_2 => "MEMORY_AND_DISK_2"
-    case MEMORY_AND_DISK_SER => "MEMORY_AND_DISK_SER"
-    case MEMORY_AND_DISK_SER_2 => "MEMORY_AND_DISK_SER_2"
-    case OFF_HEAP => "OFF_HEAP"
-    case _ => throw new IllegalArgumentException(s"Invalid StorageLevel: $level")
-  }
-
   /**
    * Get parquet file reader
    *
@@ -190,22 +133,11 @@ class Spark4_2Adapter extends BaseSpark4Adapter {
     Spark42ParquetReader.build(vectorized, sqlConf, options, hadoopConf)
   }
 
-  override def createLanceFileReader(vectorized: Boolean,
-                                     sqlConf: SQLConf,
-                                     options: Map[String, String],
-                                     hadoopConf: Configuration): Option[SparkColumnarFileReader] = {
-    Some(new SparkLanceReaderBase(vectorized))
-  }
-
   override def createVortexFileReader(vectorized: Boolean,
                                       sqlConf: SQLConf,
                                       options: Map[String, String],
                                       hadoopConf: Configuration): Option[SparkColumnarFileReader] = {
     Some(new SparkVortexReaderBase(vectorized))
-  }
-
-  override def stopSparkContext(jssc: JavaSparkContext, exitCode: Int): Unit = {
-    jssc.sc.stop(exitCode)
   }
 
   override def getDateTimeRebaseMode(): LegacyBehaviorPolicy.Value = {
@@ -218,18 +150,6 @@ class Spark4_2Adapter extends BaseSpark4Adapter {
       .map(LegacyBehaviorPolicy.withName)
     fromSqlConf.orElse(fromSparkConf)
       .getOrElse(SQLConf.get.getConf(SQLConf.PARQUET_REBASE_MODE_IN_WRITE))
-  }
-
-  override def isLegacyBehaviorPolicy(value: Object): Boolean = {
-    value == LegacyBehaviorPolicy.LEGACY
-  }
-
-  override def isTimestampNTZType(dataType: DataType): Boolean = {
-    dataType == DataTypes.TimestampNTZType
-  }
-
-  override def getRebaseSpec(policy: String): RebaseDateTime.RebaseSpec = {
-    RebaseDateTime.RebaseSpec(LegacyBehaviorPolicy.withName(policy))
   }
 
   override def isVariantProjectionStruct(structType: StructType): Boolean = {
